GithubRepoTalker/Talker/requirements.txt
openai
whisper
PyGithub
